lista-personagens

+++ src/main/java/com/ciandt/treinamento/controller/CharactersController.java
@@ -1,5 +1,7 @@
 package com.ciandt.treinamento.controller;
 
+import javax.xml.stream.events.Characters;
+
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -7,14 +9,20 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.ciandt.treinamento.service.impl.CharacterServiceImpl;
+
+
 @RestController
 @RequestMapping("api/v1")
 public class CharactersController {
 
 	@GetMapping("/characters/")
 	public ResponseEntity<Object> returnAllCharacters(
-			@RequestParam(name = "quantidade", required = false) Long quantidade) {
-//		return (ResponseEntity<Object>) characters.characterList(quantidade);
+			@RequestParam(name = "quantidade", required = false) Integer quantidade) {
+//		if (quantidade < 20)
+//			return (ResponseEntity<Object>) CharacterServiceImpl.returnAllCharacters(quantidade);
+//		else
+//			return (ResponseEntity<Object>) CharacterServiceImpl.returnAllCharacters(20);
 		return null;
 
 	}
